Made some changes based on feedback
