+++ app/build.gradle.kts
@@ -11,7 +11,7 @@ android {
     defaultConfig {
         applicationId = "com.merahit.sokoconnect"
         minSdk = 24
-        targetSdk = 36
+        targetSdk = 35
         versionCode = 1
         versionName = "1.0"
